Implement bounded contexts and lint tooling
