+++ pymerp/backend/build.gradle
@@ -1,7 +1,11 @@
+import org.gradle.api.plugins.quality.Checkstyle
+
 plugins {
   id "java"
   id "org.springframework.boot" version "3.3.3"
   id "io.spring.dependency-management" version "1.1.6"
+  id "checkstyle"
+  id "com.diffplug.spotless" version "6.25.0"
 }
 
 group = "com.datakomerz"
@@ -48,3 +52,45 @@ dependencies {
 }
 
 tasks.named("test") { useJUnitPlatform() }
+
+checkstyle {
+  toolVersion = "10.17.0"
+  configDirectory = file("config/checkstyle")
+}
+
+tasks.withType(Checkstyle).configureEach {
+  reports { report ->
+    report.xml.required.set(false)
+    report.html.required.set(true)
+  }
+}
+
+def boundedContextTargets = [
+  "src/main/java/com/datakomerz/pymes/auth/api/**/*.java",
+  "src/main/java/com/datakomerz/pymes/auth/application/**/*.java",
+  "src/main/java/com/datakomerz/pymes/customers/api/**/*.java",
+  "src/main/java/com/datakomerz/pymes/customers/application/**/*.java",
+  "src/main/java/com/datakomerz/pymes/sales/api/**/*.java",
+  "src/main/java/com/datakomerz/pymes/sales/application/**/*.java",
+  "src/main/java/com/datakomerz/pymes/common/api/**/*.java",
+  "src/main/java/com/datakomerz/pymes/core/tenancy/CompanyContextFilter.java"
+]
+
+spotless {
+  java {
+    target(boundedContextTargets)
+    removeUnusedImports()
+    trimTrailingWhitespace()
+    endWithNewline()
+  }
+}
+
+tasks.named("checkstyleMain").configure {
+  include(boundedContextTargets.collect { it.replace("src/main/java/", "") })
+}
+
+tasks.named("checkstyleTest").configure {
+  include(boundedContextTargets.collect { it.replace("src/main/java/", "") })
+}
+
+tasks.named("check") { dependsOn("spotlessCheck") }
