+++ pymerp/backend/config/checkstyle/checkstyle.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<!DOCTYPE module PUBLIC
+  "-//Checkstyle//DTD Checkstyle Configuration 1.3//EN"
+  "https://checkstyle.org/dtds/configuration_1_3.dtd">
+<module name="Checker">
+  <module name="RegexpSingleline">
+    <property name="format" value="\s+$"/>
+    <property name="message" value="Trailing whitespace is not allowed."/>
+  </module>
+  <module name="TreeWalker">
+    <module name="UnusedImports"/>
+  </module>
+</module>
