+++ pymerp/backend/src/main/java/com/datakomerz/pymes/customers/application/CustomerMapper.java
@@ -1,0 +1,22 @@
+package com.datakomerz.pymes.customers.application;
+
+import com.datakomerz.pymes.customers.Customer;
+import com.datakomerz.pymes.customers.dto.CustomerResponse;
+import org.springframework.stereotype.Component;
+
+@Component
+public class CustomerMapper {
+
+  public CustomerResponse toResponse(Customer customer) {
+    return new CustomerResponse(
+      customer.getId(),
+      customer.getName(),
+      customer.getAddress(),
+      customer.getLat(),
+      customer.getLng(),
+      customer.getPhone(),
+      customer.getEmail(),
+      customer.getSegment()
+    );
+  }
+}
